NewsBot/requirements.txt
aiohttp==3.6.2
astral==1.10.1
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
ciso8601==2.1.3
click==7.1.2
cryptography==2.9.2
decorator==4.4.2
dialogflow==1.1.0
feedparser==5.2.1
Flask==1.1.2
fuzzywuzzy==0.18.0
gnewsclient==1.12
google-api-core==1.22.1
google-auth==1.20.1
googleapis-common-protos==1.52.0
grpcio==1.31.0
gunicorn==20.0.4
homeassistant==0.114.4
idna==2.10
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
python-slugify==4.0.1
python-telegram-bot==12.8
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
ruamel.yaml==0.15.100
six==1.15.0
text-unidecode==1.3
tornado==6.0.4
urllib3==1.25.10
voluptuous==0.11.7
voluptuous-serialize==2.4.0
Werkzeug==1.0.1
yarl==1.4.2
zipp==3.1.0
